refactor(postprocess): split postprocess.c into domain TUs
Split the monolithic 1634-line postprocess.c into 6 focused
translation units, each under 360 lines:

- postprocess_init.c     (358 lines) — initialization, resize, dummy textures
- postprocess_apply.c    (360 lines) — begin/end render pass, UBO sync, compositing
- postprocess_setters.c  (268 lines) — enable/disable/toggle, setters, preset apply
- postprocess_shader.c   (223 lines) — shader cache, compile, dynamic switching
- postprocess_readback.c (199 lines) — PBO readback, histogram, luminance, matrices
- postprocess_cleanup.c  (97 lines)  — resource teardown

New internal header postprocess_internal.h provides shared enums
and pp_-prefixed internal function declarations.

Public API (postprocess.h) is unchanged.

Closes #249
