+++ CMakeLists.txt
@@ -124,9 +124,14 @@ set(SOURCES
     src/pbr.c
     src/perf_mode.c
     src/perf_timer.c
-    src/postprocess.c
+    src/postprocess_apply.c
+    src/postprocess_cleanup.c
+    src/postprocess_init.c
     src/postprocess_input.c
     src/postprocess_presets.c
+    src/postprocess_readback.c
+    src/postprocess_setters.c
+    src/postprocess_shader.c
     src/render_utils.c
     src/scene_cleanup.c
     src/scene_init.c
